# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Property Sale Screening API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/screenings": {
      "post": {
        "tags": [
          "screenings"
        ],
        "summary": "Trigger Screening",
        "description": "\u30b9\u30af\u30ea\u30fc\u30cb\u30f3\u30b0\u3092\u975e\u540c\u671f\u3067\u5b9f\u884c\u958b\u59cb\u3059\u308b\uff08202 Accepted\uff09.",
        "operationId": "trigger_screening_api_v1_screenings_post",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreeningTriggerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/screenings/{run_id}": {
      "get": {
        "tags": [
          "screenings"
        ],
        "summary": "Get Screening Status",
        "description": "\u30b9\u30af\u30ea\u30fc\u30cb\u30f3\u30b0\u5b9f\u884c\u72b6\u614b\u3092\u53d6\u5f97\u3059\u308b.",
        "operationId": "get_screening_status_api_v1_screenings__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreeningRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/kpi/effectiveness": {
      "get": {
        "tags": [
          "kpi"
        ],
        "summary": "Get Effectiveness",
        "description": "\u52b9\u679c\u691c\u8a3c KPI \u30b9\u30ca\u30c3\u30d7\u30b7\u30e7\u30c3\u30c8\u3092\u53d6\u5f97\u3059\u308b\uff08\u8aad\u307f\u53d6\u308a\u5c02\u7528\uff09.",
        "operationId": "get_effectiveness_api_v1_kpi_effectiveness_get",
        "parameters": [
          {
            "name": "period_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\u96c6\u8a08\u671f\u9593\uff08\u958b\u59cb\uff09",
              "title": "Period From"
            },
            "description": "\u96c6\u8a08\u671f\u9593\uff08\u958b\u59cb\uff09"
          },
          {
            "name": "period_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\u96c6\u8a08\u671f\u9593\uff08\u7d42\u4e86\uff09",
              "title": "Period To"
            },
            "description": "\u96c6\u8a08\u671f\u9593\uff08\u7d42\u4e86\uff09"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EffectivenessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kpi/work-logs": {
      "post": {
        "tags": [
          "kpi"
        ],
        "summary": "Create Work Log",
        "description": "\u5de5\u6570\u30ed\u30b0\u3092\u8a18\u9332\u3059\u308b\uff08JWT \u8a8d\u8a3c\u5fc5\u9808\uff09.",
        "operationId": "create_work_log_api_v1_kpi_work_logs_post",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkLogCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkLogSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "kpi"
        ],
        "summary": "List Work Logs",
        "description": "\u5de5\u6570\u30ed\u30b0\u4e00\u89a7\u3092\u53d6\u5f97\u3059\u308b.",
        "operationId": "list_work_logs_api_v1_kpi_work_logs_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\u958b\u59cb\u65e5",
              "title": "From"
            },
            "description": "\u958b\u59cb\u65e5"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\u7d42\u4e86\u65e5",
              "title": "To"
            },
            "description": "\u7d42\u4e86\u65e5"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkLogListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EffectivenessResponse": {
        "properties": {
          "snapshots": {
            "items": {
              "$ref": "#/components/schemas/KpiSnapshotSchema"
            },
            "type": "array",
            "title": "Snapshots",
            "description": "\u671f\u9593\u5185\u306e KPI \u30b9\u30ca\u30c3\u30d7\u30b7\u30e7\u30c3\u30c8\u4e00\u89a7\uff08\u6607\u9806\uff09"
          },
          "latest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KpiSnapshotSchema"
              },
              {
                "type": "null"
              }
            ],
            "description": "\u6700\u65b0\u306e KPI \u30b9\u30ca\u30c3\u30d7\u30b7\u30e7\u30c3\u30c8"
          }
        },
        "type": "object",
        "title": "EffectivenessResponse",
        "description": "\u52b9\u679c\u691c\u8a3c KPI \u30ec\u30b9\u30dd\u30f3\u30b9\uff08\u6700\u65b0 + \u63a8\u79fb\uff09."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KpiSnapshotSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "\u30b9\u30ca\u30c3\u30d7\u30b7\u30e7\u30c3\u30c8ID"
          },
          "period_from": {
            "type": "string",
            "format": "date",
            "title": "Period From",
            "description": "\u96c6\u8a08\u671f\u9593\uff08\u958b\u59cb\uff09"
          },
          "period_to": {
            "type": "string",
            "format": "date",
            "title": "Period To",
            "description": "\u96c6\u8a08\u671f\u9593\uff08\u7d42\u4e86\uff09"
          },
          "universe_coverage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Universe Coverage",
            "description": "\u30e6\u30cb\u30d0\u30fc\u30b9\u30ab\u30d0\u30ec\u30c3\u30b8\uff08%\uff09"
          },
          "traceability_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Traceability Rate",
            "description": "\u30c8\u30ec\u30fc\u30b9\u53ef\u80fd\u7387\uff08%\uff09"
          },
          "avg_structure_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Structure Score",
            "description": "\u5e73\u5747\u69cb\u9020\u30b9\u30b3\u30a2"
          },
          "reproducibility_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reproducibility Score",
            "description": "\u518d\u73fe\u6027\u30b9\u30b3\u30a2"
          },
          "total_workload_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Workload Min",
            "description": "\u7dcf\u5de5\u6570\uff08\u5206\uff09"
          },
          "workload_reduction_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workload Reduction Rate",
            "description": "\u5de5\u6570\u524a\u6e1b\u7387\uff08%\uff09"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "\u4f5c\u6210\u65e5\u6642"
          }
        },
        "type": "object",
        "required": [
          "id",
          "period_from",
          "period_to",
          "created_at"
        ],
        "title": "KpiSnapshotSchema",
        "description": "\u52b9\u679c\u691c\u8a3c KPI \u30b9\u30ca\u30c3\u30d7\u30b7\u30e7\u30c3\u30c8\uff08\u8aad\u307f\u53d6\u308a\u5c02\u7528\uff09."
      },
...
      "RunStatusEnum": {
        "type": "string",
        "enum": [
          "running",
          "success",
          "failed"
        ],
        "title": "RunStatusEnum",
        "description": "\u30b9\u30af\u30ea\u30fc\u30cb\u30f3\u30b0\u5b9f\u884c\u30b9\u30c6\u30fc\u30bf\u30b9."
      },
...
      "ScreeningRunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id",
            "description": "\u30b9\u30af\u30ea\u30fc\u30cb\u30f3\u30b0\u5b9f\u884cID"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RunStatusEnum"
              }
            ],
            "description": "\u5b9f\u884c\u30b9\u30c6\u30fc\u30bf\u30b9"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At",
            "description": "\u958b\u59cb\u65e5\u6642"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At",
            "description": "\u5b8c\u4e86\u65e5\u6642"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress",
            "description": "\u9032\u6357\u7387\uff080\u2013100 %\uff09"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "started_at",
          "progress"
        ],
        "title": "ScreeningRunResponse",
        "description": "\u30b9\u30af\u30ea\u30fc\u30cb\u30f3\u30b0\u5b9f\u884c\u72b6\u614b\u30ec\u30b9\u30dd\u30f3\u30b9."
      },
      "ScreeningTriggerResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id",
            "description": "\u767a\u884c\u3055\u308c\u305f\u30b9\u30af\u30ea\u30fc\u30cb\u30f3\u30b0\u5b9f\u884cID"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RunStatusEnum"
              }
            ],
            "description": "\u8d77\u52d5\u76f4\u5f8c\u306e\u30b9\u30c6\u30fc\u30bf\u30b9\uff08\u5e38\u306b running\uff09",
            "default": "running"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message",
            "description": "\u8d77\u52d5\u78ba\u8a8d\u30e1\u30c3\u30bb\u30fc\u30b8"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "message"
        ],
        "title": "ScreeningTriggerResponse",
        "description": "\u30b9\u30af\u30ea\u30fc\u30cb\u30f3\u30b0\u8d77\u52d5\u30ec\u30b9\u30dd\u30f3\u30b9."
      },
...
      "TaskTypeEnum": {
        "type": "string",
        "enum": [
          "primary_screening",
          "deep_dive",
          "report",
          "other"
        ],
        "title": "TaskTypeEnum",
        "description": "\u5de5\u6570\u30ed\u30b0\u306e\u30bf\u30b9\u30af\u7a2e\u5225."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkLogCreateRequest": {
        "properties": {
          "task_type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TaskTypeEnum"
              }
            ],
            "description": "\u30bf\u30b9\u30af\u7a2e\u5225"
          },
          "duration_min": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Duration Min",
            "description": "\u6240\u8981\u6642\u9593\uff08\u5206\uff09"
          },
          "screening_run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Screening Run Id",
            "description": "\u95a2\u9023\u30b9\u30af\u30ea\u30fc\u30cb\u30f3\u30b0\u5b9f\u884cID\uff08\u4efb\u610f\uff09"
          },
          "period_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Label",
            "description": "\u671f\u9593\u30e9\u30d9\u30eb\uff08\u4efb\u610f\uff09"
          },
          "logged_on": {
            "type": "string",
            "format": "date",
            "title": "Logged On",
            "description": "\u8a18\u9332\u5bfe\u8c61\u65e5"
          }
        },
        "type": "object",
        "required": [
          "task_type",
          "duration_min",
          "logged_on"
        ],
        "title": "WorkLogCreateRequest",
        "description": "\u5de5\u6570\u30ed\u30b0\u8a18\u9332\u30ea\u30af\u30a8\u30b9\u30c8."
      },
      "WorkLogListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkLogSchema"
            },
            "type": "array",
            "title": "Items",
            "description": "\u5de5\u6570\u30ed\u30b0\u4e00\u89a7"
          },
          "total_min": {
            "type": "integer",
            "title": "Total Min",
            "description": "\u5408\u8a08\u5de5\u6570\uff08\u5206\uff09",
            "default": 0
          }
        },
        "type": "object",
        "title": "WorkLogListResponse",
        "description": "\u5de5\u6570\u30ed\u30b0\u4e00\u89a7\u30ec\u30b9\u30dd\u30f3\u30b9."
      },
      "WorkLogSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "\u5de5\u6570\u30ed\u30b0ID"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "\u8a18\u9332\u8005\u30e6\u30fc\u30b6\u30fcID"
          },
          "task_type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TaskTypeEnum"
              }
            ],
            "description": "\u30bf\u30b9\u30af\u7a2e\u5225"
          },
          "duration_min": {
            "type": "integer",
            "title": "Duration Min",
            "description": "\u6240\u8981\u6642\u9593\uff08\u5206\uff09"
          },
          "screening_run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Screening Run Id",
            "description": "\u95a2\u9023\u30b9\u30af\u30ea\u30fc\u30cb\u30f3\u30b0\u5b9f\u884cID"
          },
          "period_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Label",
            "description": "\u671f\u9593\u30e9\u30d9\u30eb"
          },
          "logged_on": {
            "type": "string",
            "format": "date",
            "title": "Logged On",
            "description": "\u8a18\u9332\u5bfe\u8c61\u65e5"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "\u4f5c\u6210\u65e5\u6642"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "task_type",
          "duration_min",
          "logged_on",
          "created_at"
        ],
        "title": "WorkLogSchema",
        "description": "\u5de5\u6570\u30ed\u30b0."
      }
    }
  }
}